Entity sizes are now dynamic.
As the size of the entity may change in the future I decied to place the
entity's xPos/yPos to the middle of the AABB (Axis aligned bounding box)
so shifting entity sizes could be easily handled.

Was initially going to have a pointer to the size of the entity.
But was unable to build a AABB collision test between an entity
accessed by DP and another entity accessed by the Y index register without
constantly loading/unloading the X register.