Refactored entity code to use subclasses.
There is now a subclass of Entity called EntityPhysics that handles
the entity's velcoity and tile collisions.

Have also added a macro (ENTITY_STRUCT) that builds a .struct containing
the struct fields of the class, but leaves the .struct open so more
fields can be added later.

Each entity takes 64 bytes of memory. This is to easily allow future
expansion and if I can pad the entity table location to 64 bytes
I get a 1 in 4 odds of the ZP.l aligning to 0 to get a speedup.